Bug 1352093 - Use an integer type for sNumLivingContexts that won't overflow when counting objects. r=mstange
MozReview-Commit-ID: 2w9y0flEdrX

UltraBlame original commit: 9faaded6a48d90c21b2773c656f6b87da552b197

+++ dom/canvas/CanvasRenderingContext2D.h
@@ -578,7 +578,7 @@ protected:
 
 
 
-  static uint32_t sNumLivingContexts;
+  static uintptr_t sNumLivingContexts;
 
   static mozilla::gfx::DrawTarget* sErrorTarget;
